% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnitPost.R
\name{KnitPost}
\alias{KnitPost}
\title{simply knits .md to .rmd}
\usage{
KnitPost(input, base.url = "/")
}
\arguments{
\item{input}{file name of .Rmd}

\item{base.url}{some other thing}
}
\value{
.md into directory
}
\description{
http://jfisher-usgs.github.io/
}
\author{
j Fisher
}

