% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnitPost_AB.R
\name{KnitPost_AB}
\alias{KnitPost_AB}
\title{KnitPost by Andrew Brooks}
\usage{

  KnitPost_AB(site.path = "C:/Users/asus/Documents/GitHub/Pokemon_FieldStudies/",
  overwriteAll = F, overwriteOne = NULL)
}
\arguments{
\item{overwriteOne}{}
}
\description{
How to use KnitPost
1. Configure the directories at the top of KnitPost to match the file system of your blog.
I could have included these as parameters, but I figured I wouldn’t re-architect my blog
very often… so I left them hard-coded.
}
\details{
2. Create an R Markdown post and save it as a .Rmd file in rmd.path. Be sure to include the
proper YAML front matter for Jekyll. This tripped me up initially. I forgot to change the
date from the knitr style date (“Month, day YYYY”) that auto-generates when you create a
new .Rmd to a Jekyll style date (‘YYYY-MM-DD’).

3. Run KnitPost to publish your R Markdown file.

http://brooksandrew.github.io/simpleblog/articles/blogging-with-r-markdown-and-jekyll-using-knitr/
}
\author{
Andrew Brooks
}

