% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnitPost_JF.R
\name{KnitPost_JF}
\alias{KnitPost_JF}
\title{Simply knits .md to .rmd}
\usage{
KnitPost_JF(input, base.url = "/")
}
\arguments{
\item{input}{file name of .Rmd}

\item{base.url}{some other thing}
}
\value{
.md into directory
}
\description{
http://jfisher-usgs.github.io/
}
\author{
j Fisher
}

